LOCALISATION started
Just added es and en folders at app, and copies of page.tsx. When linked to the es or en directories, Next.js directly finds its page.tsx, and customizes its theme with the main layout.tsx and globals.css.

This provides a faster solution to setup for webs with no great amount of contents or updates. For a long term multilingual support, App Router i18n localisation, with its own json files for languages and variables to decide which images to use, is encouraged.